+++ src/cpx_io/cpx_system/cpx_ap/builder/apdd_information_builder.py
@@ -51,8 +51,13 @@ def build_actual_variant(apdd, module_code):
 def build_apdd_information(apdd, variant):
     """Builds one ApddInformation"""
 
-    product_category = apdd["Variants"]["DeviceIdentification"]["ProductCategory"]
-    product_family = apdd["Variants"]["DeviceIdentification"]["ProductFamily"]
+    product_category = apdd["Variants"]["DeviceIdentification"].get("ProductCategory")
+    product_family = apdd["Variants"]["DeviceIdentification"].get("ProductFamily")
+
+    if not product_category or not product_family:
+        raise RuntimeError(f"{variant.name} can not be build due to outdated firmware. "
+                           "Consider updating your cpx-ap module with the latest firmware "
+                           "from https://www.festo.com")
 
     return ApModule.ApddInformation(
         variant.description,