+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- CPX-AP: Outdated and not-supported module firmware will now raise RuntimeError with suggestion to update firmware.
 ### Changed
 - CPX-AP: Added diagnosis example